# cli/internal/hub/openapi.json
{
  "components": {
    "schemas": {
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "correlation_id": {
            "description": "The request's correlation id, as also returned in the X-Correlation-ID header.",
            "type": "string"
          },
          "detail": {
            "description": "Explanation specific to this occurrence.",
            "type": "string"
          },
          "errors": {
            "description": "Per-field detail, when the failure was a validation failure.",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array"
          },
          "status": {
            "description": "HTTP status code, repeated for client convenience.",
            "examples": [
              404
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "Short, static summary of the problem type.",
            "examples": [
              "Not Found"
            ],
            "type": "string"
          },
          "type": {
            "description": "A URI reference to documentation for this error type.",
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "description": "Where it is wrong, e.g. body.targets[0] or path.revision.",
            "type": "string"
          },
          "message": {
            "description": "What is wrong.",
            "type": "string"
          },
          "value": {
            "description": "The offending value, echoed back."
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
...
      "PackageCapabilities": {
        "additionalProperties": false,
        "properties": {
          "expected": {
            "items": {
              "$ref": "#/components/schemas/PackageCapability"
            },
            "type": "array"
          },
          "inferred": {
            "items": {
              "$ref": "#/components/schemas/PackageCapability"
            },
            "type": "array"
          },
          "scanned": {
            "type": "boolean"
          }
        },
        "required": [
          "scanned",
          "inferred",
          "expected"
        ],
        "type": "object"
      },
      "PackageCapability": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "description": "The scoping: hosts for network, paths for filesystem, command names for shell.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "indefinite": {
            "type": "boolean"
          },
          "level": {
            "description": "How much trust it demands. A shell capability is never below review (FR-018).",
            "enum": [
              "scoped",
              "allowlisted",
              "review"
            ],
            "examples": [
              "allowlisted"
            ],
            "type": "string"
          },
          "name": {
            "enum": [
              "network",
              "filesystem.read",
              "filesystem.write",
              "shell"
            ],
            "examples": [
              "network"
            ],
            "type": "string"
          }
        },
        "required": [
          "name",
          "level",
          "detail"
        ],
        "type": "object"
      },
      "PackageComponent": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "skill",
              "mcp",
              "ext"
            ],
            "examples": [
              "skill"
            ],
            "type": "string"
          },
          "name": {
            "examples": [
              "terraform-module-review"
            ],
            "type": "string"
          },
          "note": {
            "examples": [
              "SKILL.md + scripts/"
            ],
            "type": "string"
          },
          "path": {
            "examples": [
              "skills/terraform-module-review"
            ],
            "type": "string"
          }
        },
        "required": [
          "kind",
          "name",
          "path"
        ],
        "type": "object"
      },
      "PackageDependent": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "description": "How this profile resolves the package.",
            "enum": [
              "latest",
              "pinned",
              "range"
            ],
            "examples": [
              "pinned"
            ],
            "type": "string"
          },
          "name": {
            "examples": [
              "Platform baseline"
            ],
            "type": "string"
          },
          "range": {
            "examples": [
              "^1.2"
            ],
            "type": "string"
          },
          "slug": {
            "examples": [
              "platform-baseline"
            ],
            "type": "string"
          },
          "version": {
            "examples": [
              "1.3.0"
            ],
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "mode"
        ],
        "type": "object"
      },
      "PackageDetail": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "$ref": "#/components/schemas/PackageCapabilities"
          },
          "category": {
            "description": "The admin-curated category (FR-049). Empty when none was chosen.",
            "examples": [
              "Infrastructure"
            ],
            "type": "string"
          },
          "components": {
            "description": "Derived from the FILE TREE. No manifest field enumerates components (FR-017, R1).",
            "items": {
              "$ref": "#/components/schemas/PackageComponent"
            },
            "type": "array"
          },
          "dependents": {
            "description": "Profiles using this package that the CALLER MAY READ, and no others (FR-044).",
            "items": {
              "$ref": "#/components/schemas/PackageDependent"
            },
            "type": "array"
          },
          "description": {
            "examples": [
              "Platform guardrails, ADR authoring and service scaffolding."
            ],
            "type": "string"
          },
          "id": {
            "description": "namespace/name — the first segment of the publisher slug, not the whole slug.",
            "examples": [
              "example/platform-toolkit"
            ],
            "type": "string"
          },
          "kind": {
            "description": "Decided by which manifest is at the tree root, never by a manifest field.",
            "enum": [
              "plugin",
              "skill"
            ],
            "examples": [
              "plugin"
            ],
            "type": "string"
          },
          "manifest": {
            "type": "string"
          },
          "manifestObject": {
            "enum": [
              "plugin.json",
              "SKILL.md"
            ],
            "examples": [
              "plugin.json"
            ],
            "type": "string"
          },
          "name": {
            "description": "The manifest name. No manifest field and no column carries a human title.",
            "examples": [
              "platform-toolkit"
            ],
            "type": "string"
          },
          "origin": {
            "$ref": "#/components/schemas/PackageOrigin"
          },
          "publisher": {
            "$ref": "#/components/schemas/PackagePublisher"
          },
          "tags": {
            "description": "The latest version's manifest keywords. Tags belong to the version, not the package.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "verdict": {
            "enum": [
              "scanning",
              "clean",
              "flagged",
              "rejected"
            ],
            "examples": [
              "clean"
            ],
            "type": "string"
          },
          "version": {
            "description": "The latest visible version's semver — the one the panels below describe.",
            "examples": [
              "1.3.0"
            ],
            "type": "string"
          },
          "versions": {
            "description": "Every visible version, newest first.",
            "items": {
              "$ref": "#/components/schemas/PackageVersion"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "publisher",
          "origin",
          "version",
          "verdict",
          "tags",
          "manifestObject",
          "manifest",
          "components",
          "capabilities",
          "versions",
          "dependents"
        ],
        "type": "object"
      },
      "PackageOrigin": {
        "additionalProperties": false,
        "properties": {
          "parentId": {
            "examples": [
              "example/platform-toolkit"
            ],
            "type": "string"
          },
          "parentName": {
            "examples": [
              "platform-toolkit"
            ],
            "type": "string"
          },
          "specVersion": {
            "examples": [
              "1.0.0"
            ],
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "PackagePublisher": {
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "examples": [
              "Platform Engineering"
            ],
            "type": "string"
          },
          "slug": {
            "examples": [
              "example/platform"
            ],
            "type": "string"
          },
          "verified": {
            "description": "Set by a catalog admin and never inferred from the namespace.",
            "type": "boolean"
          }
        },
        "required": [
          "slug",
          "displayName",
          "verified"
        ],
        "type": "object"
      },
...
      "PackageVersion": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "digest": {
            "description": "Empty while a registration is still being fetched: the digest is computed on write.",
            "pattern": "^sha256:[0-9a-f]{64}$",
            "type": "string"
          },
          "distTag": {
            "enum": [
              "latest",
              "archived",
              "none"
            ],
            "examples": [
              "latest"
            ],
            "type": "string"
          },
          "objectKey": {
            "examples": [
              "skills/example/platform-toolkit/1.3.0/bundle.tar.zst"
            ],
            "type": "string"
          },
          "pinnedBy": {
            "format": "int64",
            "type": "integer"
          },
          "sizeBytes": {
            "format": "int64",
            "type": "integer"
          },
          "verdict": {
            "enum": [
              "scanning",
              "clean",
              "flagged",
              "rejected"
            ],
            "examples": [
              "clean"
            ],
            "type": "string"
          },
          "version": {
            "examples": [
              "1.3.0"
            ],
            "type": "string"
          }
        },
        "required": [
          "version",
          "distTag",
          "verdict",
          "createdAt",
          "objectKey",
          "pinnedBy"
        ],
        "type": "object"
      },
...
      "Profile": {
        "additionalProperties": false,
        "properties": {
          "headRevision": {
            "description": "The most recent published revision number.",
            "examples": [
              7
            ],
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "description": "Human-readable name as shown in the web UI.",
            "examples": [
              "Platform baseline"
            ],
            "type": "string"
          },
          "packageCount": {
            "description": "Packages in the head revision, excluding skipped entries.",
            "examples": [
              12
            ],
            "format": "int64",
            "type": "integer"
          },
          "slug": {
            "description": "URL-safe identifier, unique within the organisation.",
            "examples": [
              "platform-baseline"
            ],
            "type": "string"
          },
          "visibility": {
            "description": "How this profile came to be readable by this identity.",
            "enum": [
              "organisation",
              "shared",
              "private"
            ],
            "examples": [
              "organisation"
            ],
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "packageCount",
          "headRevision"
        ],
        "type": "object"
      },
      "ProfileList": {
        "additionalProperties": false,
        "properties": {
          "profiles": {
            "description": "Readable profiles. Order is not part of the contract.",
            "items": {
              "$ref": "#/components/schemas/Profile"
            },
            "type": "array"
          }
        },
        "required": [
          "profiles"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "opaque",
        "description": "The `access_token` issued by /v1/device/token, or a web session token, sent as `Authorization: Bearer \u003ctoken\u003e`. The token is OPAQUE — one base64url segment of random bytes, with no header, payload or claims. Do not decode one and do not read `exp` from it: a token's lifetime is the `expires_in` returned beside it.",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionMintSecret": {
        "description": "A secret shared between the `serve web` and `serve api` roles, compared in constant time. It authorises minting a session for ANY subject, so it is the most privileged credential this hub accepts. The mint is refused outright when no secret is configured: there is no default and no development bypass.",
        "in": "header",
        "name": "X-Session-Mint-Secret",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Agent Manager maintainers",
      "url": "https://github.com/WindKube/agent-manager"
    },
    "description": "Self-hosted registry for AI agent plugins and skills. This document is emitted from the huma operation definitions in internal/api (constitution principle V) and is the source every generated client is built from.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "Agent Manager",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/packages/{namespace}/{name}": {
      "get": {
        "description": "Description, origin, tags, manifest, components, capabilities, version history and dependent profiles for one package (FR-016). The path is the package id, whose two segments are the namespace and the name. Two panels are scoped to the caller: the dependent profiles are exactly the ones this identity may read (FR-044), and each version's `pinnedBy` counts only those — an unscoped count beside a scoped list would leak the existence of private profiles by arithmetic. `capabilities.scanned` distinguishes a version that was scanned and reaches nothing from one that has never been scanned; the two produce identical empty lists, and only that flag tells them apart.",
        "operationId": "getPackage",
        "parameters": [
          {
            "description": "The publishing namespace — the FIRST segment of the publisher slug, as it appears in the catalog id. Not the whole slug.",
            "in": "path",
            "name": "namespace",
            "required": true,
            "schema": {
              "description": "The publishing namespace — the FIRST segment of the publisher slug, as it appears in the catalog id. Not the whole slug.",
              "examples": [
                "example"
              ],
              "type": "string"
            }
          },
          {
            "description": "The package name within that namespace.",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "description": "The package name within that namespace.",
              "examples": [
                "platform-toolkit"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageDetail"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No usable session. The caller must sign in; there is no anonymous view."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No such package, or it has no published version."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request could not be completed."
          }
        },
        "summary": "One package's detail",
        "tags": [
          "catalog"
        ]
      }
    },
    "/v1/profiles": {
      "get": {
        "description": "Returns exactly the profiles this identity may read via direct membership or group mapping, and no others (FR-044). Not a filtered view of a larger list — unreadable profiles are not enumerated at all.",
        "operationId": "listProfiles",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileList"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Missing, expired or invalid token."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request could not be completed."
          }
        },
        "summary": "Profiles readable by this identity",
        "tags": [
          "profiles"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}